Prevent multiple MLLP adapter instances from submitting metrics with the same label
When running multiple instances of MLLP adapter in the same VM instance, we could get the following error: "CreateTimeSeries failed: rpc error: code = InvalidArgument desc = One or more TimeSeries could not be written: Points must be written in order. One or more of the points specified had an older start time than the most recent point.: timeSeries[...]". This is because multiple MLLP adapter instances tried to submit metrics with the same label. Adding UUID to the label to prevent the possible collision.

PiperOrigin-RevId: 325509636